Move conformance helpers to internal/conformance
Move GetTestCases() and WriteMetrics() from the root hegel package to
internal/conformance. Update all cmd/conformance/*/main.go binaries to
import from the new location. Delete conformance.go and conformance_test.go
from root; equivalent tests live in internal/conformance/helpers_test.go.
